+++ CHANGES.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Option '--erasing' renamed into '--mapping'.
 - The export option `--requiring` does not require as argument a file with extension `.v` anymore: the argument must be a module name.
 
 ## 2.5.1 (2024-07-22)